fix issue

+++ src/common/models/oidc_user_meta.go
@@ -6,7 +6,7 @@ import (
 
 type OIDCUserMetaData struct {
 	ID           int64     `orm:"pk;auto;column(id)" json:"id"`
-	UserID       int64     `orm:"column(user_id)" json:"user_id"`
+	UserID       int       `orm:"column(user_id)" json:"user_id"`
 	Name         string    `orm:"column(name)" json:"name"`
 	Value        string    `orm:"column(value)" json:"value"`
 	CreationTime time.Time `orm:"column(creation_time);auto_now_add" json:"creation_time"`
